requirements.txt
streamlit
pandas
plotly.express
snowflake-connector-python